# docs/openapi/ingest_api_v1.yaml
openapi: 3.0.3
info:
  title: Fractsoul Ingest API
  version: 1.1.0
  description: >
    API de ingesta y lectura de telemetria para el MVP de operacion de granja ASIC.
    La autenticacion por API key es configurable via entorno (`API_AUTH_ENABLED`, `API_KEYS`).
servers:
  - url: http://localhost:8080
security:
  - ApiKeyAuth: []
tags:
  - name: health
  - name: telemetry
  - name: efficiency
  - name: anomaly
paths:
  /healthz:
    get:
      tags: [health]
      summary: Healthcheck del servicio
      security: []
      responses:
        '200':
          description: Servicio operativo
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /v1/telemetry/ingest:
    post:
      tags: [telemetry]
      summary: Ingesta un evento de telemetria
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestRequest'
      responses:
        '202':
          description: Evento aceptado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
        '400':
          description: Payload invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: API key faltante/invalida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Payload supera limite configurado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Timestamp fuera de rango
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/telemetry/readings:
    get:
      tags: [telemetry]
      summary: Lista lecturas de telemetria
      parameters:
        - $ref: '#/components/parameters/site_id'
        - $ref: '#/components/parameters/rack_id'
        - $ref: '#/components/parameters/miner_id'
        - $ref: '#/components/parameters/model'
        - $ref: '#/components/parameters/status'
        - $ref: '#/components/parameters/from'
        - $ref: '#/components/parameters/to'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Lecturas obtenidas
          content:
            application/json:
              schema:
                type: object
                properties:
                  request_id:
                    type: string
                    format: uuid
                  count:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/TelemetryReading'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /v1/telemetry/summary:
    get:
      tags: [telemetry]
      summary: Resumen agregado de telemetria
      parameters:
        - $ref: '#/components/parameters/site_id'
        - $ref: '#/components/parameters/rack_id'
        - $ref: '#/components/parameters/miner_id'
        - $ref: '#/components/parameters/model'
        - name: window_minutes
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1440
            default: 60
      responses:
        '200':
          description: Resumen obtenido
          content:
            application/json:
              schema:
                type: object
                properties:
                  request_id:
                    type: string
                    format: uuid
                  summary:
                    $ref: '#/components/schemas/TelemetrySummary'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /v1/telemetry/sites/{site_id}/racks/{rack_id}/readings:
    get:
      tags: [telemetry]
      summary: Lecturas por sitio/rack
      parameters:
        - name: site_id
          in: path
          required: true
          schema:
            type: string
        - name: rack_id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/miner_id'
        - $ref: '#/components/parameters/model'
        - $ref: '#/components/parameters/status'
        - $ref: '#/components/parameters/from'
        - $ref: '#/components/parameters/to'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Lecturas obtenidas
          content:
            application/json:
              schema:
                type: object
                properties:
                  request_id:
                    type: string
                    format: uuid
                  site_id:
                    type: string
                  rack_id:
                    type: string
                  count:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/TelemetryReading'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /v1/telemetry/miners/{miner_id}/timeseries:
    get:
      tags: [telemetry]
      summary: Serie temporal agregada por maquina
      parameters:
        - name: miner_id
          in: path
          required: true
          schema:
            type: string
        - name: resolution
          in: query
          schema:
            type: string
            enum: [minute, hour]
            default: minute
        - $ref: '#/components/parameters/from'
        - $ref: '#/components/parameters/to'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Serie temporal obtenida
          content:
            application/json:
              schema:
                type: object
                properties:
                  request_id:
                    type: string
                    format: uuid
                  miner_id:
                    type: string
                  resolution:
                    type: string
                    enum: [minute, hour]
                  from:
                    type: string
                    format: date-time
                  to:
                    type: string
                    format: date-time
                  count:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/MinerSeriesPoint'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /v1/efficiency/miners:
    get:
      tags: [efficiency]
      summary: Eficiencia J/TH por equipo
      parameters:
        - $ref: '#/components/parameters/site_id'
        - $ref: '#/components/parameters/rack_id'
        - $ref: '#/components/parameters/miner_id'
        - $ref: '#/components/parameters/model'
        - $ref: '#/components/parameters/window_minutes'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Eficiencia por equipo obtenida
          content:
            application/json:
              schema:
                type: object
                properties:
                  request_id:
                    type: string
                    format: uuid
                  count:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/MinerEfficiency'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /v1/efficiency/racks:
    get:
      tags: [efficiency]
      summary: Eficiencia J/TH agregada por rack
      parameters:
        - $ref: '#/components/parameters/site_id'
        - $ref: '#/components/parameters/rack_id'
        - $ref: '#/components/parameters/model'
        - $ref: '#/components/parameters/window_minutes'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Eficiencia por rack obtenida
          content:
            application/json:
              schema:
                type: object
                properties:
                  request_id:
                    type: string
                    format: uuid
                  count:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/RackEfficiency'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /v1/efficiency/sites:
    get:
      tags: [efficiency]
      summary: Eficiencia J/TH agregada por sitio
      parameters:
        - $ref: '#/components/parameters/site_id'
        - $ref: '#/components/parameters/model'
        - $ref: '#/components/parameters/window_minutes'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Eficiencia por sitio obtenida
          content:
            application/json:
              schema:
                type: object
                properties:
                  request_id:
                    type: string
                    format: uuid
                  count:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/SiteEfficiency'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /v1/anomalies/miners/{miner_id}/analyze:
    get:
      tags: [anomaly]
      summary: Analisis de anomalias para un equipo
      parameters:
        - name: miner_id
          in: path
          required: true
          schema:
            type: string
        - name: resolution
          in: query
          schema:
            type: string
            enum: [minute, hour]
            default: minute
        - $ref: '#/components/parameters/from'
        - $ref: '#/components/parameters/to'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/ambient_c'
      responses:
        '200':
          description: Reporte de anomalias generado
          content:
            application/json:
              schema:
                type: object
                properties:
                  request_id:
                    type: string
                    format: uuid
                  summary_line:
                    type: string
                  report:
                    $ref: '#/components/schemas/AnomalyReport'
        '401':
          $ref: '#/components/responses/Unauthorized'

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  parameters:
    site_id:
      name: site_id
      in: query
      schema:
        type: string
    rack_id:
      name: rack_id
      in: query
      schema:
        type: string
    miner_id:
      name: miner_id
      in: query
      schema:
        type: string
    model:
      name: model
      in: query
      schema:
        type: string
    status:
      name: status
      in: query
      schema:
        type: string
        enum: [ok, warning, critical, offline]
    from:
      name: from
      in: query
      schema:
        type: string
        format: date-time
    to:
      name: to
      in: query
      schema:
        type: string
        format: date-time
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
    window_minutes:
      name: window_minutes
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 1440
        default: 60
    ambient_c:
      name: ambient_c
      in: query
      schema:
        type: number

  responses:
    Unauthorized:
      description: API key faltante o invalida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    IngestRequest:
      type: object
      required: [event_id, timestamp, site_id, rack_id, miner_id, metrics]
      properties:
        event_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        site_id:
          type: string
        rack_id:
          type: string
        miner_id:
          type: string
        firmware_version:
          type: string
        metrics:
          type: object
          required: [hashrate_ths, power_watts, temp_celsius, fan_rpm, efficiency_jth, status]
          properties:
            hashrate_ths: { type: number }
            power_watts: { type: number }
            temp_celsius: { type: number }
            fan_rpm: { type: integer }
            efficiency_jth: { type: number }
            status:
              type: string
              enum: [ok, warning, critical, offline]
        tags:
          type: object
          additionalProperties:
            type: string

    IngestResponse:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
        accepted:
          type: boolean
        event_id:
          type: string
          format: uuid
        queue_topic:
          type: string
        ingested_at:
          type: string
          format: date-time

    TelemetryReading:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        event_id: { type: string, format: uuid }
        site_id: { type: string }
        rack_id: { type: string }
        miner_id: { type: string }
        miner_model: { type: string }
        firmware_version: { type: string }
        hashrate_ths: { type: number }
        power_watts: { type: number }
        temp_celsius: { type: number }
        fan_rpm: { type: integer }
        efficiency_jth: { type: number }
        status: { type: string, enum: [ok, warning, critical, offline] }
        load_pct: { type: number }
        tags:
          type: object
          additionalProperties:
            type: string
        ingested_at:
          type: string
          format: date-time

    TelemetrySummary:
      type: object
      properties:
        window_minutes: { type: integer }
        samples: { type: integer }
        avg_hashrate_ths: { type: number }
        avg_power_watts: { type: number }
        avg_temp_celsius: { type: number }
        p95_temp_celsius: { type: number }
        max_temp_celsius: { type: number }
        avg_fan_rpm: { type: number }
        avg_efficiency_jth: { type: number }

    MinerSeriesPoint:
      type: object
      properties:
        bucket: { type: string, format: date-time }
        samples: { type: integer }
        avg_hashrate_ths: { type: number }
        avg_power_watts: { type: number }
        avg_temp_celsius: { type: number }
        max_temp_celsius: { type: number }
        avg_fan_rpm: { type: number }
        avg_efficiency_jth: { type: number }
        critical_events: { type: integer }

    MinerEfficiency:
      type: object
      properties:
        site_id: { type: string }
        rack_id: { type: string }
        miner_id: { type: string }
        miner_model: { type: string }
        samples: { type: integer }
        avg_hashrate_ths: { type: number }
        avg_power_watts: { type: number }
        avg_temp_celsius: { type: number }
        avg_ambient_celsius: { type: number }
        raw_jth: { type: number }
        compensated_jth: { type: number }
        baseline_jth: { type: number }
        delta_pct: { type: number }
        thermal_band: { type: string }
        last_seen_at: { type: string, format: date-time }

    RackEfficiency:
      type: object
      properties:
        site_id: { type: string }
        rack_id: { type: string }
        miners: { type: integer }
        samples: { type: integer }
        avg_hashrate_ths: { type: number }
        avg_power_watts: { type: number }
        avg_temp_celsius: { type: number }
        avg_ambient_celsius: { type: number }
        raw_jth: { type: number }
        compensated_jth: { type: number }
        last_seen_at: { type: string, format: date-time }

    SiteEfficiency:
      type: object
      properties:
        site_id: { type: string }
        miners: { type: integer }
        racks: { type: integer }
        samples: { type: integer }
        avg_hashrate_ths: { type: number }
        avg_power_watts: { type: number }
        avg_temp_celsius: { type: number }
        avg_ambient_celsius: { type: number }
        raw_jth: { type: number }
        compensated_jth: { type: number }
        last_seen_at: { type: string, format: date-time }

    AnomalyFeatureVector:
      type: object
      properties:
        samples: { type: integer }
        avg_hashrate_ths: { type: number }
        avg_power_watts: { type: number }
        avg_temp_celsius: { type: number }
        max_temp_celsius: { type: number }
        avg_fan_rpm: { type: number }
        avg_efficiency_jth: { type: number }
        hashrate_drop_pct: { type: number }
        hashrate_trend_ths_per_hour: { type: number }
        temp_trend_c_per_hour: { type: number }
        power_trend_w_per_hour: { type: number }
        ambient_celsius: { type: number }
        compensated_jth: { type: number }
        thermal_band: { type: string }

    AnomalyDetection:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        triggered: { type: boolean }
        score: { type: number }
        evidence:
          type: object
          additionalProperties:
            type: number

    AnomalyRecommendation:
      type: object
      properties:
        parameter: { type: string }
        suggested_delta: { type: string }
        reason: { type: string }

    AnomalyReport:
      type: object
      properties:
        miner_id: { type: string }
        site_id: { type: string }
        rack_id: { type: string }
        miner_model: { type: string }
        window_from: { type: string, format: date-time }
        window_to: { type: string, format: date-time }
        features:
          $ref: '#/components/schemas/AnomalyFeatureVector'
        hotspot:
          $ref: '#/components/schemas/AnomalyDetection'
        hash_degradation:
          $ref: '#/components/schemas/AnomalyDetection'
        severity_score: { type: number }
        severity_label: { type: string }
        probable_cause: { type: string }
        recommendations:
          type: array
          items:
            $ref: '#/components/schemas/AnomalyRecommendation'

    ErrorResponse:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
        code:
          type: string
        message:
          type: string
        details:
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
              items:
                type: object
                additionalProperties: true
            - type: string